Compress migrations
This commit reduces the number of migrations that were created during
this feature's development (which is a no-brainer, as this is the first
feature branch).

In the upcoming development steps, this should be done before merging
into the actual development branch.

Additionally a minor typo in the stylesheet sources was fixed.