Adjust LTC field on Sample model
Ok, this is a quite small commit code-wise, but it will determine the
future handling of large numbers semantically.

In the context of the app, those large numbers happen most prominently
while dealing with coin-values of the game, as those go into the
quadrillions and quintillions.

Main idea is this: The values are stored in Django's ``Decimal`` field,
following the convention of assuming a magnitude of *trillions*.